Update tests and accept new baselines.
1. Better coverage.
2. It turns out that, `() => this.a` need to stay legal, no matter what
the context is, because `this` always refers to an outer this, never a
contextually typed `this`. I do not try to type a captured `this` any
better than it currently is -- and for object types this means it stays
`any`. (Methods are already correct and function bodies are correct as of
this PR, so it's not a big hole.)