Initial version of inference for this types.
It has a lot of duplication and I'm not sure it's right. Specifically, I
don't think that the callee can be an inference target, only a source of
inference. Even I know that it couldn't be, I'm not sure yet how to change
the code to ensure this -- I just copied the existing code.