Simplify quickInfo for this expressions.
Previously it was 'class X' because it directly referred to the containing
class' symbol. Now it's just 'X' because it instead uses the default code
that finds the *type* by location instead of the *symbol*. This allows the code to
work for non-class types.

The correct change may be to modify interface handling in
getSymbolDisplayPartsDocumentationAndSymbolKind, but at this point I want
to get all tests passing before I make a change that affects non-this
types.